move constants out of internal
